requirements.txt
fastapi
uvicorn
python-dotenv
openai
supabase
sqlalchemy
psycopg2-binary
